Launch SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

PS3 - Added material green icons for SU

Omni-Change-Id: I2af95a1a4396101ff87448cd0ae20addf235af52

Conflicts:
	res/xml/dashboard_categories.xml
	src/com/android/settings/SettingsActivity.java

My addition:

Don't show Superuser option if SuperSU is installed

This will make sure it is hidden if a user switches from superuser to SuperSU after enabling superuser

PS2: Pre-emptive line lenght fixe

PS3: Update slim_strings for latest changes

Change-Id: I719fe2054639c73118d662cc8a1f6db72fbb0896

Conflicts:
	res/values/slim_strings.xml

Conflicts:
	res/values/slim_strings.xml